rewrite back to int, g(20000, 200),40/17, s=2.34

+++ src/Graph.java
@@ -3,7 +3,7 @@ import java.util.*;
 // Represents a directed graph
 public class Graph {
 
-    public record Edge(int source, int destination, Double weight) {}
+    public record Edge(int source, int destination, int weight) {}
 
     public final int V;
     public final int E;
